requirements.txt
Flask==3.0.2
langchain==0.1.5
openai==1.11.1
PyJWT==2.8.0
python-dotenv==1.0.1
